Refactor to Feature-based clean Architecture
