+++ step_ai/lib/features/personal/presentation/widgets/no_knowledge_panel.dart
