+++ step_ai/lib/shared/widgets/app_name_widget.dart
