+++ step_ai/lib/shared/widgets/chat_bar.dart
